fix declaration bug

+++ xv6-public/defs.h
@@ -129,8 +129,8 @@ int             change_process_queue(int, int);
 float           calc_process_bjf_rank(struct proc* p);
 int             uncle_count(int);
 int             find_process_lifetime(int);
-int             set_bjf_process(int,float,float,float);
-void            set_bjf_system(float,float,float);
+int             set_bjf_process(int,int,int,int);
+void            set_bjf_system(int,int,int);
 
 
 // swtch.S
